Add mailchimp commands to command folder
