+++ tests/cypress/support/commands.js
@@ -25,7 +25,7 @@
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
-import './mailchimpApi/requests';
+import './commands/mailchimp';
 
 const state = {};
